time now measured in milliseconds (adapted rest of code)

+++ src/utilities_solution.c
@@ -21,7 +21,7 @@ void solve_with_method(instance *inst, solution *sol) {
     allocate_solution(sol, inst->nnodes);
     _mkdir("./results"); //, 0777);
 
-    int timelimit = inst->timelimit - get_elapsed_time(inst->t_start);
+    int timelimit =(int) inst->timelimit - get_elapsed_time(inst->t_start);
 
     if (strcmp(inst->asked_method, NEAREST_NEIGHBOR) == 0) {
 
@@ -34,7 +34,7 @@ void solve_with_method(instance *inst, solution *sol) {
         nearest_neighbor(inst, sol, 0);
         if (inst->param1 != 3 && inst->param1 != 5) { inst->param1 = DEAULT_K; }
         if (inst->param2 < 1) { inst->param2 = DEFAULT_REPS; }
-        int elapsed_time = get_elapsed_time(inst->t_start);
+        double elapsed_time = get_elapsed_time(inst->t_start);
         vns(inst, sol, (timelimit-elapsed_time), inst->param1, inst->param2);
         
     } else if (strcmp(inst->asked_method, TABU_SEARCH) == 0) {
